+++ Project.toml
@@ -3,12 +3,14 @@ uuid = "5078a376-72f3-5289-bfd5-ec5146d43c02"
 version = "0.13"
 
 [deps]
+ArrayLayouts = "4c555306-a7a7-4459-81d9-ec55ddd5c99a"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
+ArrayLayouts = "0.1"
 FillArrays = "0.7,0.8"
 MacroTools = "0.4.5,0.5"
 StaticArrays = "0.8,0.9,0.10,0.11"
